pyproject.toml
[project]
name = "document360-mcp"
version = "0.1.0"
description = "Document360 MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.7.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]

[project.scripts]
document360-mcp = "main:main"